Add button to always power new structs
Extend the functionality of powerOnNewStruct to handle both electricity and region support. Return true iff the new building is powered on.

To allow region support structure detection, add new field "s_type" to structures that consume support and have ON/OFF buttons. At this time, the field is not added to structures without ON/OFF buttons, such as Swarm Satellite.

Resolve incongruities in existing structure definitions:
- Add support() function to the zoo
- Check for available support before enabling Tau Ceti mining ships and whaling ships